test: 파워매니저 테스트 파워의 범위를 체크하는 메서드 추가
